+++ include/RendererAPI.hpp
@@ -40,7 +40,7 @@ namespace Ra
         virtual void SetClearColor(const glm::vec4& color) = 0;
         virtual void Clear() = 0;
 
-        virtual void DrawIndexed(const Ref<VertexArray>& array, DrawMode mode, std::uint32_t indexCount) = 0;
+        virtual void DrawIndexed(const Ref<VertexArray>& array, DrawMode mode, std::size_t indexCount) = 0;
 
         static Scope<RendererAPI> Create();
